pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "centinela"
version = "0.1.0"
description = "Centinela — Agente IA autónomo con seguridad de grado empresarial"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "rypcloud"},
]

dependencies = [
    # LLM & Agent Orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-aws>=0.2.0",
    "boto3>=1.35.0",
    # API Gateway
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # CLI
    "typer>=0.15.0",
    "rich>=13.0.0",
    # Config & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    # Vector DB & Memory
    "qdrant-client>=1.12.0",
    # Sandbox
    "docker>=7.0.0",
    # Auth
    "pyjwt>=2.9.0",
    # Logging
    "structlog>=24.0.0",
    # Web tools
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    # Messaging
    "python-telegram-bot>=22.0",
    "slack-bolt>=1.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
centinela = "centinela.interfaces.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/centinela"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
